% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{annotateByCellnamePattern}
\alias{annotateByCellnamePattern}
\title{Creates a meta data entry with name group.name for the given annotation}
\usage{
annotateByCellnamePattern(
  obj.in,
  group.name,
  annotation,
  order = NULL,
  use.base = NULL
)
}
\arguments{
\item{obj.in}{Seurat object to annotate}

\item{group.name}{new group name}

\item{annotation}{annotation}

\item{order}{whether the annotation (which will be stored as factor)}

\item{use.base}{}
}
\description{
Creates a meta data entry with name group.name for the given annotation
}
