% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cellIDForClusters}
\alias{cellIDForClusters}
\title{cellIDForClusters selects cells given a value to look for and a meta-data column to search in}
\usage{
cellIDForClusters(obj.in, targetVar, clusters)
}
\arguments{
\item{obj.in}{Seurat object}

\item{targetVar}{column to look in}

\item{clusters}{values to look in for}
}
\description{
cellIDForClusters selects cells given a value to look for and a meta-data column to search in
}
