% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combine_plot_grid}
\alias{combine_plot_grid}
\title{Combines several ggplot2 objects}
\usage{
combine_plot_grid(...)
}
\arguments{
\item{...}{ggplot2 objects}
}
\value{
summarized ggplot2 object
}
\description{
Combines several ggplot2 objects with cowplot and also stores its data
}
