% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combine_plot_grid_list}
\alias{combine_plot_grid_list}
\title{Combines several ggplot2 objects}
\usage{
combine_plot_grid_list(plotlist, ...)
}
\arguments{
\item{plotlist}{ggplot2 objects}

\item{...}{additional parameters forwarded to cowplot's plot_grid}
}
\value{
summarized ggplot2 object
}
\description{
Combines several ggplot2 objects with cowplot and also stores its data
}
