% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{compareCellsByCluster}
\alias{compareCellsByCluster}
\title{Splits all Seurat objects in the input list by a specific meta tag; useful for sample integration}
\usage{
compareCellsByCluster(
  inobj,
  cellsID1,
  cellsID2,
  suffix1,
  suffix2,
  group.by = NULL,
  assay = "RNA",
  test = "t",
  outfolder = "./",
  fcCutoff = 0.25,
  heatmap.plot = FALSE,
  heatmap.pcutoff = 0.05,
  heatmap.addgenes = NULL
)
}
\arguments{
\item{inobj}{Seurat object}

\item{cellsID1}{cell names of cells to include for group1}

\item{cellsID2}{cell names of cells to include for group2}

\item{suffix1}{name of group1}

\item{suffix2}{name of group2}

\item{group.by}{meta-data column to perform the comparisons for}

\item{assay}{assay to use for comparison (should be RNA)}

\item{test}{test to perform for differential expression}

\item{outfolder}{output folder/stump where to store the results}

\item{fcCutoff}{minimal logFC to report}

\item{heatmap.plot}{whether to plot a heatmap for the comparison}

\item{heatmap.pcutoff}{p-value cut off for genes to show in the heatmap}

\item{heatmap.addgenes}{additional genes to include in heatmap}
}
\value{
list of differential expression data frame
}
\description{
Splits all Seurat objects in the input list by a specific meta tag; useful for sample integration
}
