% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{compareClusters}
\alias{compareClusters}
\title{Performs a comparison between cellsID1 and cellsID2}
\usage{
compareClusters(
  scdata,
  cellsID1,
  cellsID2,
  suffix1,
  suffix2,
  prefix = "cluster",
  test = "MAST",
  assay = "RNA",
  outfolder = "./",
  fcCutoff = 0.25,
  heatmap.plot = FALSE,
  heatmap.pcutoff = 0.05,
  heatmap.addgenes = NULL
)
}
\arguments{
\item{cellsID1}{cell names of cells to include for group1}

\item{cellsID2}{cell names of cells to include for group2}

\item{suffix1}{name of group1}

\item{suffix2}{name of group2}

\item{test}{test to perform for differential expression}

\item{assay}{assay to use for comparison (should be RNA)}

\item{outfolder}{output folder/stump where to store the results}

\item{fcCutoff}{minimal logFC to report}

\item{heatmap.plot}{whether to plot a heatmap for the comparison}

\item{heatmap.pcutoff}{p-value cut off for genes to show in the heatmap}

\item{heatmap.addgenes}{additional genes to include in heatmap}

\item{inobj}{Seurat object}
}
\value{
differential expression data frame
}
\description{
Performs a comparison between cellsID1 and cellsID2
}
