% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{converts a list of mouse gene symbols to human gene symbols}
\usage{
convertHumanGeneList(x)
}
\arguments{
\item{x}{mouse gene symbols}
}
\value{
human gene symbols
}
\description{
converts a list of mouse gene symbols to human gene symbols
}
