% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{converts a list of human gene symbols to mouse gene symbols}
\usage{
convertMouseGeneList(x)
}
\arguments{
\item{x}{human gene symbols}
}
\value{
mouse gene symbols
}
\description{
converts a list of human gene symbols to mouse gene symbols
}
