% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{createDirIfNotThere}
\alias{createDirIfNotThere}
\title{createDirIfNotThere - creates directory if it does not exist yet}
\usage{
createDirIfNotThere(infolder)
}
\arguments{
\item{which}{directory to create}
}
\description{
createDirIfNotThere - creates directory if it does not exist yet
}
