% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{enhancedHeatMap}
\alias{enhancedHeatMap}
\title{Creates an enhanced Dot Plot}
\usage{
enhancedHeatMap(
  obj.in,
  plot_gois,
  group.by = "idents",
  include_all_clusters = FALSE,
  title = NULL,
  scale.by = "GLOBAL",
  scale.limits = NULL,
  log.expression = TRUE
)
}
\arguments{
\item{obj.in}{Seurat object for plotting}

\item{plot_gois}{list of list(clusters,genes) where each entry defines the cluster to be shown, and genes the corresponding genes}

\item{group.by}{name of the meta.data column used for grouping cells}

\item{include_all_clusters}{whether to fill up the shown groups with the remaining ones from the category}

\item{title}{title of the plot}

\item{scale.by}{how to show/scale the expression values (NULL; ALL; GLOBAL)}

\item{scale.limits}{lower and upper bounds of the scale for gene expression}

\item{log.expression}{whether to log1p the expression values (only applies to GLOBAL and ALL scale)}
}
\value{
ggplot2
}
\description{
Creates an enhanced DotPlot of the RNA-assay expression
}
