% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{getExprData_getPop}
\alias{getExprData_getPop}
\title{Calculates expression values for a set of cells}
\usage{
getExprData_getPop(
  markerObj,
  markerCells,
  sampleSuffix,
  slot = "data",
  assay = "RNA"
)
}
\arguments{
\item{markerObj}{Seurat object}

\item{markerCells}{cells for which gene expression data is to be calculated}

\item{sampleSuffix}{suffix of the column names}

\item{slot}{data slot to pull expression values from}

\item{assay}{the assay from which to pull the expression values}
}
\value{
list of summarised expression data
}
\description{
Calculates expression values for a set of cells
}
