% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{makeDEResults}
\alias{makeDEResults}
\title{Performs Marker analysis for all clusters defined by the group.by grouping. Analysis is performed in the given assay using the given test.}
\usage{
makeDEResults(inobj, group.by = NULL, assay = "SCT", test = "wilcox")
}
\arguments{
\item{inobj}{Seurat object}

\item{group.by}{column in the meta.data to perform the analysis on}

\item{assay}{assay on which the analysis is performed on}

\item{test}{the test used to perform for finding marker genes}
}
\value{
data frame with marker genes for each group/cluster
}
\description{
Performs Marker analysis for all clusters defined by the group.by grouping. Analysis is performed in the given assay using the given test.
}
