% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{makeEnrichmentPlots}
\alias{makeEnrichmentPlots}
\title{makeEnrichmentPlots - makes plots for the enrichment results generated by performEnrichtmentAnalysis}
\usage{
makeEnrichmentPlots(enrichResult, outfolder, qvalue_threshold = 0.2)
}
\arguments{
\item{enrichResult}{list of enrichment results as generated by performEnrichtmentAnalysis}

\item{outfolder}{folder where to store plots from enrichment analysis}

\item{qvalue_threshold}{qvalue threshold up to where to accept gene sets}
}
\description{
makeEnrichmentPlots - makes plots for the enrichment results generated by performEnrichtmentAnalysis
}
