% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{makeQCPlots}
\alias{makeQCPlots}
\title{Creates QC Plots for a Seurat objects and a given reduction}
\usage{
makeQCPlots(inobj, outfolder, reduction = "umap")
}
\arguments{
\item{inobj}{Seurat object}

\item{outfolder}{Output path/stump for the plots}

\item{reduction}{using which reduction for the plots}
}
\value{
Seurat object
}
\description{
Creates QC Plots for a Seurat objects and a given reduction
}
