% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{makesummary_getPop}
\alias{makesummary_getPop}
\title{Creates a summary of a vector of expression values}
\usage{
makesummary_getPop(a, suffix)
}
\arguments{
\item{a}{vector of expression values}

\item{suffix}{column suffix}
}
\value{
list of summarised expression data
}
\description{
Creates a summary of a vector of expression values
}
