% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeUMAPPlot}
\alias{makeUMAPPlot}
\title{makeUMAPPlot}
\usage{
makeUMAPPlot(
  obj.in,
  dim1,
  dim2,
  reduction = "umap",
  xmin = -11,
  xmax = 15,
  ymin = -15,
  ymax = 10.5,
  group.by = "cellnames",
  downsample = FALSE
)
}
\arguments{
\item{obj.in}{Seurat object}

\item{dim1}{meta-column for y dimension}

\item{dim2}{meta-column for x dimension}

\item{reduction}{dimensional reduction to plot}

\item{xmin}{left x border}

\item{xmax}{right x border}

\item{ymin}{lower y border}

\item{ymax}{upper y border}

\item{group.by}{which meta-column to use for coloring the cells}

\item{downsample}{whether the number of cells per group should be down-sampled to the lower number of cells per group}
}
\description{
makeUMAPPlot
}
