% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{makeVolcanos}
\alias{makeVolcanos}
\title{creates volcano plots for DE results list}
\usage{
makeVolcanos(
  loMG,
  titlePrefix,
  outname,
  restrict_labels = NULL,
  turnExpression = F,
  colors = list(neg = list(sig = "#448CCA", nosig = "#B4D1E9"), pos = list(sig =
    "#F47B78", nosig = "#FACAC9")),
  FCcutoff = 0.5,
  pCutoff = 0.05,
  highlightGene = NULL
)
}
\arguments{
\item{loMG}{list of dataframes from differential expression testing}

\item{titlePrefix}{Prefix of the title of the plots}

\item{outname}{path-stump for output (path + prefix of outname)}

\item{restrict_labels}{list of genes for each data frame to plot}

\item{turnExpression}{multiply logFC values by -1}

\item{colors}{list of list (neg(sig, nosig), pos(sig, nosig)) for colors}

\item{FCcutoff}{minimal logFC to include in plot}

\item{pCutoff}{minimal adjusted p-value to include in plot}

\item{highlightGene}{features to always include in the volcano plots}
}
\value{
None
}
\description{
creates volcano plots for DE results list
}
