% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{mergeRNAHTO}
\alias{mergeRNAHTO}
\title{Merges the list with Seurat objects and the list with HTO Seurat objects}
\usage{
mergeRNAHTO(objlist, htolist)
}
\arguments{
\item{objlist}{list of Seurat objects}

\item{htolist}{list of Seurat objects with HTO assay}
}
\value{
list of Seurat objects
}
\description{
Merges the list with Seurat objects and the list with HTO Seurat objects
}
