% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{patternList_human}
\alias{patternList_human}
\title{Returns a patternlist to identify mitochondrial and ribosomal genes in human}
\usage{
patternList_human()
}
\value{
patternlist
}
\description{
Returns a patternlist to identify mitochondrial and ribosomal genes in human
}
