% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{patternList_mouse}
\alias{patternList_mouse}
\title{Returns a patternlist to identify mitochondrial and ribosomal genes in mouse}
\usage{
patternList_mouse()
}
\value{
patternlist
}
\description{
Returns a patternlist to identify mitochondrial and ribosomal genes in mouse
}
