% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{performEnrichtmentAnalysis}
\alias{performEnrichtmentAnalysis}
\title{performEnrichtmentAnalysis - performs enrichment analysis on differential expression results provided by compareCellsByCluster}
\usage{
performEnrichtmentAnalysis(
  scobj,
  deList,
  organismName,
  outFile,
  reverseLogFC = FALSE
)
}
\arguments{
\item{scobj}{Seurat object with active assay from which the DE results are (used to get gene universe for set enrichment)}

\item{deList}{differential expression results provided by compareCellsByCluster}

\item{organismName}{human or mouse}

\item{outFile}{where to store the enrichment results as Rds file}

\item{reverseLogFC}{set true if you want to reverse logFC values (logFC=-logFC)}
}
\value{
list with enrichment results for each comparison given in deList
}
\description{
performEnrichtmentAnalysis - performs enrichment analysis on differential expression results provided by compareCellsByCluster
}
