% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{performIntegration}
\alias{performIntegration}
\title{This method performs the integration of all Seurat objects in \code{objlist} into a single Seurat object}
\usage{
performIntegration(
  objlist,
  intname,
  features.integration = 3000,
  gex.method.normalization = "LogNormalize",
  gex.assay = "RNA",
  gex.runpca = TRUE,
  gex.dims = 30,
  gex.method.integration = "rpca",
  gex.k.filter = 200,
  gex.k.anchor = 5,
  gex.k.weight = 100,
  add.do = FALSE,
  add.assay = "ABA",
  add.dims = 10,
  add.method.integration = "rpca",
  add.k.filter = 200,
  add.k.anchor = 5,
  add.features.integration = 100,
  add.k.weight = 5,
  run.parallel = TRUE
)
}
\arguments{
\item{objlist}{list of Seurat objects}

\item{intname}{output path for all plots}

\item{features.integration}{Features to perform the gene expression integration on}

\item{gex.method.normalization}{Which normalization strategy to use for normalization (SCT or LogNormalize)}

\item{gex.assay}{assay to use for integration}

\item{gex.runpca}{whether to run PCA on the gene expression assay prior to integration}

\item{gex.dims}{number of dimensions for finding gene expression integration}

\item{gex.method.integration}{Which integration method to use (SCTransform, cca or rpca) for the gene expression data}

\item{gex.k.filter}{How many neighbors (k) to use when filtering anchors for the gene expression data}

\item{gex.k.anchor}{How many neighbors (k) to use when picking anchors for the gene expression data}

\item{gex.k.weight}{Number of neighbors to consider when weighting anchors for the gene expression data}

\item{add.do}{Whether an additional assay should be used for integration (e.g. CITE)}

\item{add.assay}{Which assay to use as additional assay}

\item{add.dims}{number of dimensions for finding additional integration}

\item{add.method.integration}{Which integration method to use (SCTransform, cca or rpca) for the additional data}

\item{add.k.filter}{How many neighbors (k) to use when filtering anchors for the additional data}

\item{add.k.anchor}{How many neighbors (k) to use when picking anchors for the additional data}

\item{add.features.integration}{Features to perform the additional data integration on}

\item{add.k.weight}{Number of neighbors to consider when weighting anchors for the additional data}

\item{run.parallel}{whether to run ScaleData sequentially}
}
\value{
list(integrated, multimodal) where integrated is the gene expression only integrated version, and multimodal the gex+add version
}
\description{
This method performs the integration of all Seurat objects in \code{objlist} into a single Seurat object
}
