% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{prepareIntegration}
\alias{prepareIntegration}
\title{This function prepares an integration on the Seurat objects by running NormalizeData, FindVariableFeatures per objects, SelectIntegrationFeatures on all, and then scoring cell cycle genes, running ScaleData and PCA again on all objects.}
\usage{
prepareIntegration(
  finalList,
  cc.use.genes,
  nfeatures.variable = 3000,
  nfeatures.scale = 3000,
  scale.regress = c("percent.rp", "percent.mt", "nCount_RNA", "S.Score", "G2M.Score"),
  normalize = TRUE,
  findvariable = TRUE,
  run.parallel = TRUE
)
}
\arguments{
\item{finalList}{list of Seurat objects}

\item{cc.use.genes}{list of s.genes and g2m.genes}

\item{nfeatures.variable}{number of variable features to detect}

\item{nfeatures.scale}{number of features to scale}

\item{scale.regress}{which variables to regress out during ScaleData}

\item{normalize}{whether to normalize the data}

\item{findvariable}{whether to run FindVariableFeatures}

\item{run.parallel}{whether scale data should run in parallel. if FALSE, ScaleData and RunPCA will be run in a sequential context.}
}
\value{
list of data=list of Seurat objects and features=features determined by SelectIntegrationFeatures
}
\description{
This function prepares an integration on the Seurat objects by running NormalizeData, FindVariableFeatures per objects, SelectIntegrationFeatures on all, and then scoring cell cycle genes, running ScaleData and PCA again on all objects.
}
