% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{preprocessIntegrated}
\alias{preprocessIntegrated}
\title{Preprocesses an integrated Seurat object for downstream analysis}
\usage{
preprocessIntegrated(
  obj.in,
  useAssay,
  inname,
  do.scale = T,
  num.pcs = 50,
  resolution = 0.5,
  plot.reduction = "umap",
  dim.reduction = "umap",
  with.hto = TRUE,
  run.parallel = TRUE,
  run.umap_neighbors = TRUE,
  clusters.graph.name = NULL
)
}
\arguments{
\item{obj.in}{Seurat object}

\item{useAssay}{which assay to use}

\item{inname}{output folder of all plots}

\item{do.scale}{whether to scale the object}

\item{num.pcs}{number of PCs to calculate and use for UMAP/neighbors}

\item{resolution}{resolution for the clustering}

\item{plot.reduction}{which reduction to use for plotting}

\item{dim.reduction}{name of the reduction in which the PCA will be stored, and which is used for UMAP and Neighbors}

\item{with.hto}{whether also HTO plots should be prepared}

\item{run.parallel}{whether the ScaleData function should run in parallel or sequential}

\item{run.umap_neighbors}{whether to run RunUMAP, FindNeighbors and FindClusters}
}
\value{
preprocessed Seurat object
}
\description{
Preprocesses an integrated Seurat object for downstream analysis
}
