% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{processCITE}
\alias{processCITE}
\title{Takes a list of antibody capture feature-cell-matrices, the list of Seurat objects and a list of relevant hashtag oligo IDs for each sample}
\usage{
processCITE(objlist, imats, assayName = "ADT", HTOs = NULL)
}
\arguments{
\item{objlist}{list of Seurat objects}

\item{imats}{list of antibody capture feature-cell-matrices}

\item{assayName}{name of the newly created CITE-seq assay}

\item{HTOs}{hashtag oligo IDs to ignore for processing CITE}
}
\value{
list of Seurat objects with HTO assay
}
\description{
Takes a list of antibody capture feature-cell-matrices, the list of Seurat objects and a list of relevant hashtag oligo IDs for each sample
}
