% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{processHTO}
\alias{processHTO}
\title{Takes a list of antibody capture feature-cell-matrices, the list of Seurat objects and a list of relevant hashtag oligo IDs for each sample}
\usage{
processHTO(inputMatrices, objlist, relevantHTOs)
}
\arguments{
\item{inputMatrices}{list of antibody capture feature-cell-matrices}

\item{objlist}{list of Seurat objects}

\item{relevantHTOs}{relevant hashtag oligo IDs for each sample}
}
\value{
list of Seurat objects with HTO assay
}
\description{
Takes a list of antibody capture feature-cell-matrices, the list of Seurat objects and a list of relevant hashtag oligo IDs for each sample
}
