% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pValueVlnBoxPlot}
\alias{pValueVlnBoxPlot}
\title{Creates a combined Violin- and Box-Plot with split violins and significance testing}
\usage{
pValueVlnBoxPlot(
  obj.sc,
  feature,
  group.by,
  split.by,
  split.values,
  dsrCols,
  onelineLabel = FALSE,
  pt.size = 0
)
}
\arguments{
\item{obj.sc}{Seurat object for plotting}

\item{feature}{genes to plot}

\item{group.by}{name of the meta.data column used for grouping cells}

\item{split.by}{name of the meta.data column used for dodging violins}

\item{dsrCols}{named list of colors}

\item{onelineLabel}{significance-test results in one line}

\item{pt.size}{size of the single value dots}

\item{min.threshold}{minimum number of cells in group}
}
\value{
ggplot2 object
}
\description{
Creates a combined Violin- and Box-Plot with split violins and significance testing
}
