% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{qcFilterHTO}
\alias{qcFilterHTO}
\title{Plot quality control plots for the Seurat objects with HTO assay.}
\usage{
qcFilterHTO(htoObjList)
}
\arguments{
\item{objlist}{list of Seurat objects with HTO assay}
}
\value{
list of Seurat objects with HTO assay
}
\description{
Plot quality control plots for the Seurat objects with HTO assay.
}
