% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{readH5Files}
\alias{readH5Files}
\title{Reads in a list of mtx files and returns two named lists of gene expression counts and Antibody Capture Counts}
\usage{
readH5Files(
  files,
  sample_element = 3,
  sample_processor = function(x) {
     return(x)
 }
)
}
\arguments{
\item{files}{list of mtx paths}

\item{sample_element}{in the split path, which element is the sample name}

\item{sample_processor}{function modifying sample name}
}
\value{
list of gex and ab (gene expression, antibody capture) lists of per cell counts
}
\description{
Reads in a list of mtx files and returns two named lists of gene expression counts and Antibody Capture Counts
}
