% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_plot}
\alias{save_plot}
\title{Saves a ggplot object to png, pdf and svg}
\usage{
save_plot(plotobj, outname, fig.width, fig.height, save.data = TRUE)
}
\arguments{
\item{plotobj}{ggplot2 object to plot}

\item{outname}{path (without prefix) to where the plot should be stored, e.g. folder1/plot1}

\item{fig.width}{width of the plot}

\item{fig.height}{height of the plot}

\item{save.data}{whether to also save the data}
}
\value{
None
}
\description{
Also saves a data file containing all data relevant to reproduce the plot
}
