% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{scatterAndFilter}
\alias{scatterAndFilter}
\title{Takes a list of seurat objects and creates QC plots and filters all objects according to nFeature_RNA, nCount_RNA and percent.mt}
\usage{
scatterAndFilter(
  objlist,
  nfeature_rna.lower = 100,
  nfeature_rna.upper = 6000,
  ncount_rna.lower = 500,
  percent_mt.upper = 7
)
}
\arguments{
\item{objlist}{list of Seurat objects}

\item{nfeature_rna.lower}{lower bound of nFeature_RNA}

\item{nfeature_rna.upper}{upper bound of nFeature_RNA}

\item{ncount_rna.lower}{lower bound of nCount_RNA}

\item{percent_mt.upper}{upper bound of percent.mt}
}
\value{
filtered objlist
}
\description{
Takes a list of seurat objects and creates QC plots and filters all objects according to nFeature_RNA, nCount_RNA and percent.mt
}
