% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{splitObjListByGroup}
\alias{splitObjListByGroup}
\title{Splits all Seurat objects in the input list by a specific meta tag; useful for sample integration}
\usage{
splitObjListByGroup(objlist, group.by)
}
\arguments{
\item{objlist}{list of Seurat objects}

\item{group.by}{meta-column on which values the Seurat object should be split}
}
\value{
list of Seurat objects
}
\description{
Splits all Seurat objects in the input list by a specific meta tag; useful for sample integration
}
