% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{SplitVlnBoxPlot}
\alias{SplitVlnBoxPlot}
\title{Creates a combined Violin- and Box-Plot with split violins}
\usage{
SplitVlnBoxPlot(
  obj.sc,
  gene,
  group.by = "idents",
  split.by = NULL,
  pt.size = 0,
  assay = "RNA",
  min.threshold = 3,
  col = NULL,
  per.sample = TRUE
)
}
\arguments{
\item{obj.sc}{Seurat object for plotting}

\item{gene}{genes to plot}

\item{group.by}{name of the meta.data column used for grouping cells}

\item{split.by}{name of the meta.data column used for dodging violins}

\item{pt.size}{size of the single value dots}

\item{assay}{assay to use for plotting}

\item{min.threshold}{minimum number of cells in group}

\item{col}{named list of colors}

\item{per.sample}{performs the min.threshold check per split-by-group}
}
\value{
ggplot2
}
\description{
Creates a combined Violin- and Box-Plot with split violins
}
