% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{toObjList}
\alias{toObjList}
\title{Transforms a list of gene expression matrices into Seurat objects, normalizes data and calculate variable features}
\usage{
toObjList(inputMatrices, patternlist, variable.features = 3000)
}
\arguments{
\item{inputMatrices}{list of count matrices to transform into Seurat object}

\item{patternlist}{}

\item{variable.features}{list of mtx paths}
}
\value{
list of gex and ab (gene expression, antibody capture) lists of per cell counts
}
\description{
Transforms a list of gene expression matrices into Seurat objects, normalizes data and calculate variable features
}
